requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
spotipy==2.23.0
python-dotenv==1.0.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
umap-learn==0.5.5
python-multipart==0.0.6
httpx==0.25.2 